changes for ios fixes

+++ plugin/ios/Plugin/Marker.swift
@@ -4,7 +4,7 @@ import Capacitor
 public struct Marker {
     let coordinate: LatLng
     let opacity: Float?
-    let rotation: Float?
+    let rotation: Double?
     let title: String?
     let snippet: String?
     let isFlat: Bool?
@@ -56,7 +56,7 @@ public struct Marker {
 
         self.coordinate = LatLng(lat: lat, lng: lng)
         self.opacity = fromJSObject["opacity"] as? Float
-        self.rotation = fromJSObject["rotation"] as? Float
+        self.rotation = fromJSObject["rotation"] as? Double
         self.title = fromJSObject["title"] as? String
         self.snippet = fromJSObject["snippet"] as? String
         self.isFlat = fromJSObject["isFlat"] as? Bool
